edit

+++ 0x00-python_variable_annotations/7-to_kv.py
@@ -9,4 +9,4 @@ from typing import Union
 def to_kv(k: str, v: Union[int, float]) -> tuple:
     '''returns a tuple
     '''
-    return (k, v*v)
+    return (k, float(v*v))
